requirements.txt
telethon==1.36.0
python-telegram-bot==20.7